More BGBCC tweaks, working on vmem stability
More minor tweaks to BGBCC.
A lot of effort trying to get virtual memory in TestKern to work better and hopefully be some semblance of "stable". Still not entirely good, but seemingly semi-stable again after a bit of an instability crap storm...

+++ tk_qsrc/tk_clib/vxcore/include/sys/socket.h
@@ -53,7 +53,8 @@ SHUT_WR
 #endif
 
 typedef short		sa_family_t;
-typedef size_t		socklen_t;
+// typedef size_t		socklen_t;
+typedef int		socklen_t;
 
 struct sockaddr_storage {
 	sa_family_t		ss_family;
